requirements.txt
requests>=2.28.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.0.0
aiohttp>=3.8.0
json5>=0.9.0
feedparser>=6.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0